Add "nullable" modifier to functions
Functions may be specified as nullable - they may return a value, or not.
Code generators are free to ignore this modifier, or to implement
null-safe handling - for example by wrapping the result in an Option type.

Example:
service MyService {
  nullable string getName(1: i32 id)
}

The function shows that getName may return an empty result, if no user with
this id is found in the database.